swagger done

+++ controllers/studentCon.js
@@ -2,10 +2,10 @@ import db from "../config/db.js";
 import ResponseMdl from "../models/responseMdl.js";
 import log from "../utils/log.js";
 const tableName = 'Students'
-const getStudents = async () => {
+const getStudents = async (userId) => {
     log.info('StudentsCon', 'getStudents');
     const responseMdl = new ResponseMdl();
-    var data = await db.query(`SELECT * FROM ${tableName}`);
+    var data = await db.query(`SELECT * FROM ${tableName} WHERE userId=?`, userId);
     if (!data || data.length === 0) {
         responseMdl.msg = 'No data found.';
         responseMdl.status = false;
@@ -17,16 +17,16 @@ const getStudents = async () => {
 }
 
 
-const getStudent = async (id) => {
+const getStudent = async (payload) => {
     log.info('StudentsCon', 'getStudent');
     const responseMdl = new ResponseMdl();
-    var data = await db.query(`SELECT * FROM ${tableName} WHERE id=?`, id);
+    var data = await db.query(`SELECT * FROM ${tableName} WHERE userId=? AND id=?`, [payload.userId, payload.id]);
     if (!data || data.length === 0) {
         responseMdl.msg = 'No data found.';
         responseMdl.status = false;
     } else
         if (data[0].length === 0) {
-            responseMdl.msg = `Invalid student id: ${id}`;
+            responseMdl.msg = `Invalid student id or user id`;
             responseMdl.status = false;
         } else {
             responseMdl.data = data[0][0];
@@ -39,7 +39,7 @@ const getStudent = async (id) => {
 const addStudents = async (payload) => {
     log.info('StudentsCon', 'addStudents');
     const responseMdl = new ResponseMdl();
-    var data = await db.query(`INSERT INTO ${tableName} (name, age) VALUES (?,?)`, [payload.name, payload.age]);
+    var data = await db.query(`INSERT INTO ${tableName} (name, age,userId) VALUES (?,?,?)`, [payload.name, payload.age, payload.userId]);
     if (data[0].insertId == 0) {
         responseMdl.msg = 'Error'
         responseMdl.status = false;
@@ -54,7 +54,7 @@ const addStudents = async (payload) => {
 const updateStudent = async (payload) => {
     log.info('StudentsCon', 'updateStudents');
     const responseMdl = new ResponseMdl();
-    var data = await db.query(`UPDATE ${tableName} SET name = ?, age = ? WHERE id=?`, [payload.name, payload.age, payload.id]);
+    var data = await db.query(`UPDATE ${tableName} SET name = ?, age = ? WHERE userId=? AND id=?`, [payload.name, payload.age, payload.userId, payload.id]);
     if (data[0].affectedRows == 0) {
         responseMdl.msg = `Invalid student id: ${payload.id}`;
         responseMdl.status = false;
@@ -65,15 +65,15 @@ const updateStudent = async (payload) => {
     return responseMdl;
 }
 
-const deleteStudents = async (id) => {
+const deleteStudents = async (payload) => {
     log.info('StudentsCon', 'deleteStudents');
     const responseMdl = new ResponseMdl();
-    var data = await db.query(`DELETE FROM ${tableName} WHERE id=?`, id);
+    var data = await db.query(`DELETE FROM ${tableName} WHERE userId=? AND id=?`, [payload.userId, payload.id]);
     if (data[0].affectedRows == 0) {
-        responseMdl.msg = `Invalid student id: ${id}`;
+        responseMdl.msg = `Invalid student id: ${payload.id}`;
         responseMdl.status = false;
     } else {
-        responseMdl.data = { id: id };
+        responseMdl.data = { id: payload.id };
         responseMdl.msg = 'Data deleted successfully.';
     }
     return responseMdl;
